+++ Sources/VCore/Deprecations.swift
@@ -4,5 +4,3 @@
 //
 //  Created by Vakhtang Kontridze on 10/9/21.
 //
-
-import Foundation